OS-indept inner loader
