+++ patternsleuth/src/lib.rs
@@ -244,352 +244,6 @@ impl<'a, S: std::fmt::Debug + PartialEq> ScanResult<'a, S> {
     }
 }
 
-#[cfg(target_os="windows")]
-pub struct Image<'data> {
-    pub base_address: usize,
-    pub exception_directory_range: Range<usize>,
-    pub exception_children_cache: HashMap<usize, Vec<RuntimeFunction>>,
-    pub memory: Memory<'data>,
-    pub symbols: Option<HashMap<usize, String>>,
-    pub imports: HashMap<String, HashMap<String, usize>>,
-}
-#[cfg(target_os="windows")]
-impl<'data> Image<'data> {
-    fn read_inner<'memory, P: AsRef<Path>>(
-        exe_path: Option<P>,
-        load_functions: bool,
-        memory: Memory<'memory>,
-        object: object::File,
-    ) -> Result<Image<'memory>> {
-        let base_address = object.relative_address_base() as usize;
-
-        #[allow(unused_variables)]
-        let symbols = if let Some(exe_path) = exe_path {
-            #[cfg(not(feature = "symbols"))]
-            unreachable!();
-            #[cfg(feature = "symbols")]
-            {
-                let pdb_path = exe_path.as_ref().with_extension("pdb");
-                pdb_path
-                    .exists()
-                    .then(|| symbols::dump_pdb_symbols(pdb_path, base_address))
-                    .transpose()?
-            }
-        } else {
-            None
-        };
-
-        let get_ex_dir = || -> Result<Range<usize>> {
-            Ok(match object {
-                object::File::Pe64(ref inner) => {
-                    let exception_directory = inner
-                        .data_directory(object::pe::IMAGE_DIRECTORY_ENTRY_EXCEPTION)
-                        .context("no exception directory")?;
-
-                    let (address, size) = exception_directory.address_range();
-                    base_address + address as usize..base_address + (address + size) as usize
-                }
-                _ => bail!("not a PE file"),
-            })
-        };
-
-        let get_imports = || -> Result<_> {
-            Ok(match object {
-                object::File::Pe64(ref inner) => {
-                    use object::pe::ImageNtHeaders64;
-                    use object::read::pe::ImageThunkData;
-                    use object::LittleEndian as LE;
-
-                    let mut imports: HashMap<String, HashMap<String, usize>> = Default::default();
-
-                    let import_table = inner.import_table()?.unwrap();
-                    let mut import_descs = import_table.descriptors()?;
-
-                    while let Some(import_desc) = import_descs.next()? {
-                        let mut cur = HashMap::new();
-
-                        let Ok(lib_name) = import_table.name(import_desc.name.get(LE)) else {
-                            continue;
-                        };
-                        let lib_name = std::str::from_utf8(lib_name)?.to_ascii_lowercase();
-                        let mut thunks =
-                            import_table.thunks(import_desc.original_first_thunk.get(LE))?;
-                        let mut address = base_address + import_desc.first_thunk.get(LE) as usize;
-                        while let Some(thunk) = thunks.next::<ImageNtHeaders64>()? {
-                            if let Ok((_hint, name)) = import_table.hint_name(thunk.address()) {
-                                cur.insert(std::str::from_utf8(name)?.to_owned(), address);
-                                address += 8;
-                            }
-                        }
-                        imports.insert(lib_name, cur);
-                    }
-                    imports
-                }
-                _ => bail!("not a PE file"),
-            })
-        };
-
-        let mut new = Image {
-            base_address,
-            exception_directory_range: get_ex_dir().unwrap_or_default(),
-            exception_children_cache: Default::default(),
-            memory,
-            symbols,
-            imports: get_imports().unwrap_or_default(),
-        };
-
-        if load_functions {
-            new.populate_exception_cache()?;
-        }
-        Ok(new)
-    }
-    fn populate_exception_cache(&mut self) -> Result<(), MemoryAccessError> {
-        for i in self.exception_directory_range.clone().step_by(12) {
-            let f = RuntimeFunction::read(&self.memory, self.base_address, i)?;
-            self.exception_children_cache.insert(f.range.start, vec![]);
-
-            let Ok(section) = self.memory.get_section_containing(f.unwind) else {
-                // TODO disabled cause spammy
-                //println!("invalid unwind info addr {:x}", f.unwind);
-                continue;
-            };
-
-            let mut unwind = f.unwind;
-            let has_chain_info = section.section.index(unwind)? >> 3 == 0x4;
-            if has_chain_info {
-                let unwind_code_count = section.section.index(unwind + 2)?;
-
-                unwind += 4 + 2 * unwind_code_count as usize;
-                if unwind % 4 != 0 {
-                    // align
-                    unwind += 2;
-                }
-
-                if section.address() + section.data().len() > unwind + 12 {
-                    let chained = RuntimeFunction::read(section, self.base_address, unwind)?;
-
-                    // TODO disabled because it spams the log too much
-                    //let referenced = self.get_function(chained.range.start);
-
-                    //assert_eq!(Some(&chained), referenced.as_ref());
-                    //if Some(&chained) != referenced.as_ref() {
-                    //println!("mismatch {:x?} {referenced:x?}", Some(&chained));
-                    //}
-
-                    self.exception_children_cache
-                        .entry(chained.range.start)
-                        .or_default()
-                        .push(f);
-                } else {
-                    println!("invalid unwind addr {:x}", unwind);
-                }
-            }
-        }
-
-        //println!("{:#x?}", self.exception_children_cache);
-        Ok(())
-    }
-    /// Get function containing `address` from the exception directory
-    pub fn get_function(
-        &self,
-        address: usize,
-    ) -> Result<Option<RuntimeFunction>, MemoryAccessError> {
-        let size = 12;
-        let mut min = 0;
-        let mut max = self.exception_directory_range.len() / size - 1;
-
-        while min <= max {
-            let i = (max + min) / 2;
-            let addr = i * size + self.exception_directory_range.start;
-
-            let addr_begin = self.base_address + self.memory.u32_le(addr)? as usize;
-            if addr_begin <= address {
-                let addr_end = self.base_address + self.memory.u32_le(addr + 4)? as usize;
-                if addr_end > address {
-                    let unwind = self.base_address + self.memory.u32_le(addr + 8)? as usize;
-
-                    return Ok(Some(RuntimeFunction {
-                        range: addr_begin..addr_end,
-                        unwind,
-                    }));
-                } else {
-                    min = i + 1;
-                }
-            } else {
-                max = i - 1;
-            }
-        }
-        Ok(None)
-    }
-    /// Get root function containing `address` from the exception directory. This can be used to
-    /// find the start address of a function given an address in the body.
-    pub fn get_root_function(
-        &self,
-        address: usize,
-    ) -> Result<Option<RuntimeFunction>, MemoryAccessError> {
-        if let Some(f) = self.get_function(address)? {
-            let mut f = RuntimeFunction {
-                range: f.range,
-                unwind: f.unwind,
-            };
-
-            loop {
-                let mut unwind_addr = f.unwind;
-
-                let section = self.memory.get_section_containing(unwind_addr)?;
-
-                let has_chain_info = section.section.index(unwind_addr)? >> 3 == 0x4;
-                if has_chain_info {
-                    let unwind_code_count = section.section.index(unwind_addr + 2)?;
-
-                    unwind_addr += 4 + 2 * unwind_code_count as usize;
-                    if unwind_addr % 4 != 0 {
-                        // align
-                        unwind_addr += 2;
-                    }
-
-                    if section.address() + section.data().len() > unwind_addr + 12 {
-                        f = RuntimeFunction::read(section, self.base_address, unwind_addr)?;
-                    } else {
-                        todo!("not adding chain info {unwind_addr}");
-                    }
-                } else {
-                    return Ok(Some(f));
-                }
-            }
-        } else {
-            Ok(None)
-        }
-    }
-    /// Recursively get all child functions of `address` (exact). This is pulled from
-    /// `exception_children_cache` so will be empty if it has not been populated.
-    pub fn get_child_functions(
-        &self,
-        address: usize,
-    ) -> Result<Vec<RuntimeFunction>, MemoryAccessError> {
-        let mut queue = vec![address];
-        let mut all_children = vec![self.get_function(address)?.unwrap()];
-        while let Some(next) = queue.pop() {
-            if let Some(children) = self.exception_children_cache.get(&next) {
-                for child in children {
-                    queue.push(child.range().start);
-                    all_children.push(child.clone());
-                }
-            }
-        }
-        Ok(all_children)
-    }
-}
-
-
-#[cfg(target_os="linux")]
-use gimli::{BaseAddresses, EhFrame, NativeEndian};
-#[cfg(target_os="xxx")]
-impl<'data> Image<'data> {
-    fn read_inner<'memory, P: AsRef<Path>>(
-        exe_path: Option<P>,
-        load_functions: bool,
-        memory: Memory<'memory>,
-        object: object::File,
-    ) -> Result<Image<'memory>> {
-        print!("Readinner called\n");
-        let base_address = object.relative_address_base() as usize;
-        let eh_frame = object.section_by_name(".eh_frame").unwrap();
-        let eh_frame_hdr = object.section_by_name(".eh_frame_hdr").unwrap();
-        let text = object.section_by_name(".text").unwrap();
-        let bases = BaseAddresses::default()
-                    .set_eh_frame_hdr(eh_frame_hdr.address() as _)
-                    .set_eh_frame(eh_frame.address())
-                    .set_text(text.address() as _);
-        let eh_frameparsed = EhFrame::new(
-            eh_frame.data().unwrap(), 
-            NativeEndian
-        );
-        let mut entries = eh_frameparsed.entries(&bases);
-
-        let mut result = Vec::<Range<usize>>::new();
-        let mut syms = HashMap::default();
-        
-        while let Some(entry) = entries.next().context("Iter over entry failed")? {
-            match entry {
-                CieOrFde::Fde(partial) => {
-                    let fde = partial
-                        .parse(&mut EhFrame::cie_from_offset)
-                        .context("Failed parse fde item")?;
-                    // right now it's real address
-                    let start = fde.initial_address() as usize;
-                    let len = fde.len() as usize;
-                    result.push(start .. (start + len));
-                    syms.insert(start, format!("sub_{}", start));
-                }
-                CieOrFde::Cie(_) => {},
-            }
-        }
-        result.sort_by(|a,b| a.start.cmp(&b.start));
-        Ok(Image {
-            base_address: base_address,
-            memory: memory,
-            functions: Some(result),
-            symbols: Some(syms),
-            exception_children_cache: HashMap::default(),
-        })
-    }
-
-    pub fn get_function(
-        &self,
-        address: usize,
-    ) -> Result<Option<RuntimeFunction>, MemoryAccessError> {
-        self.get_root_function(address)
-    }
-
-    pub fn get_root_function(
-        &self,
-        address: usize,
-    ) -> Result<Option<RuntimeFunction>, MemoryAccessError> {
-        let x = self.functions.as_ref().unwrap();
-        // eprintln!("Finding {:#08x} within {} functions", address, x.len());
-        /*
-        let idx = x.binary_search_by(|p| match p.start.cmp(&address) {
-            Ordering::Equal => Ordering::Greater,
-            ord => ord,
-        }).unwrap_or(x.len()); 
-
-        if idx >= x.len() {
-            Ok(None)
-        } else {
-            let range = &x[idx];
-            if range.contains(&address) {
-                Ok(Some(RuntimeFunction {
-                    range: range.clone(),
-                    unwind: 0
-                }))
-            } else {
-                Ok(None)
-            }
-        }
-        */
-        x.iter().find(|p| p.contains(&address)).map(|a| 
-            Ok(Some(
-                RuntimeFunction {
-                    range: a.clone(),
-                    unwind: 0
-                }
-            ))
-        ).unwrap_or(Ok(None))
-    }
-
-    pub fn get_child_functions(
-        &self,
-        address: usize,
-    ) -> Result<Vec<RuntimeFunction>, MemoryAccessError> {
-        match self.get_function(address) {
-            Ok(Some(f)) => Ok(vec![f]),
-            Ok(None) => Ok(vec![]),
-            Err(e) => Err(e)
-        }
-    }
-}
-
 #[derive(Debug, Clone, PartialEq)]
 pub struct RuntimeFunction {
     pub range: Range<usize>,
